feat(fieldConfiguration): drop defaultValue for radio, switch and checkbox inputTypes
See withthegrid/platform#912, we are not storing default values anymore and instead display in the UI that the input is on the default, in a way that is distinctive from having that default value as hard input, which is also possible. These three inputTypes don't have a simple way to show that distinction. As default values are not so relevant for them anyway, we will drop them and (in the UI at least) enforce that one of the values is chosen (ie. undefined/ null is not possible). By default, they will be on their first option.